+++ Dau.Core/Domain/BankAccount/AccountInformationParameterTranslation.cs
@@ -6,8 +6,8 @@ namespace Dau.Core.Domain.BankAccount
 {
     public partial class AccountInformationParameterTranslation : BaseEntity
     {
-        public int AccountInfoNonTransId { get; set; }
-        public int LanguageId { get; set; }
+        public long AccountInfoNonTransId { get; set; }
+        public long LanguageId { get; set; }
         public string Parameter { get; set; }
 
         public AccountInformationParameter AccountInfoNonTrans { get; set; }
